#7: Compile time import for transform * using two classes for implicit - common extended by Spark specific one
