+++ spark2-commons/src/main/scala/za/co/absa/spark/commons/adapters/TransformAdapter.scala
@@ -16,22 +16,15 @@
 
 package za.co.absa.spark.commons.adapters
 
-import org.apache.spark.SPARK_VERSION
 import org.apache.spark.sql.Column
-import za.co.absa.commons.version.Version
-import za.co.absa.commons.version.Version.VersionStringInterpolator
 
-trait HofsAdapter {
+trait TransformAdapter {
 
   /**
    * For Spark versions prior 3.0.0, delegates to {{{hofs.transform()}}}
-   * Otherwise delegates to the native Spark method.
    */
   val transform: (Column, Column => Column) => Column = {
-    if (Version.asSemVer(SPARK_VERSION) < semver"3.0.0")
-        za.co.absa.spark.hofs.transform(_: Column, _: Column => Column)
-      else
-        org.apache.spark.sql.functions.transform(_: Column, _: Column => Column)
+    za.co.absa.spark.hofs.transform(_: Column, _: Column => Column)
   }
 
 }
